put test_provider.py in the right place
